requirements.txt
functions-framework==3.*
google-cloud-bigquery